Mojo::mysql::Pubsub - don't use REPLACE
If REPLACE finds an existing row, it deletes it and inserts a new one.
Because the timestamp for the new row defaults to "current_timestamp",
this delete/insert has the effect of just updating the timestamp for the
given pid/channel combination.  Can accomplish the same thing just by
doing INSERT ... ON DUPLICATE KEY UPDATE, but unlike REPLACE this does
not chew up a bunch of autoincrement IDs by continually creating new
rows when all it needs to do is update the timestamp.  It may also be
less prone to deadlocking, not sure.