Make Mojo::mysql::PubSub less susceptible to deadlocks/timeouts (#80)
* Mojo::mysql::Pubsub - retry once if DB error

The queries that hit the pub/sub tables are prone to deadlocking,
especially e.g. if many Minion workers are running.  After encountering
one of those errors, the pub/sub logic generally gets itself into a CPU
spin that it can't recover from.  Retrying the query in this situation
allows things to gracefully continue running.

* Mojo::mysql::Pubsub - don't use REPLACE

If REPLACE finds an existing row, it deletes it and inserts a new one.
Because the timestamp for the new row defaults to "current_timestamp",
this delete/insert has the effect of just updating the timestamp for the
given pid/channel combination.  Can accomplish the same thing just by
doing INSERT ... ON DUPLICATE KEY UPDATE, but unlike REPLACE this does
not chew up a bunch of autoincrement IDs by continually creating new
rows when all it needs to do is update the timestamp.  It may also be
less prone to deadlocking, not sure.

* Mojo::mysql::Pubsub - bigger auto_increment

Use a bigger autoincrement column since these tables can see a lot of
activity.  No need for the complexity of creating a new migration for
this - new users will pick up the new column definition and presumably
existing users are happy with it the way it already was.

* Fix .gitignore for vim backup files

* Mojo::mysql::PubSub - no need for wantarray

Mojo::mysql::Database::query only ever returns a single scalar.

* Mojo::mysql::PubSub - croak instead of die

* Mojo::mysql::PubSub - include SQL in retry warning

* Mojo::mysql::PubSub - revert auto_increment changes

* Mojo::mysql::PubSub - make number of retries configurable

* Mojo::mysql::PubSub - perltidy

Co-authored-by: Larry Leszczynski <larryl@cpan.org>